codec: Add support for decoding byte slices, even if encoded as a regular array.
This requires us checking currentEncodedType in the stream first, before
deciding how to decode the value when decoding into a slice.

There's a slight performance hit, but correctness is more important.

Also, put in flag to "conditionally" enable encoding uint as positive fixint,
and update tests so they work across binc and msgpack and regardless of flag.
(This means ensuring that all unsigned int values in test are > 128).
(I probably will not enable this flag by default, but it's there ... for placeholding and reminder).

Fixes issue #26 .

Updates issue #24 .