codec: error handling cleanup, and unsafe build tag.
This package understands the 'unsafe' tag, to allow using unsafe semantics:

  - When decoding into a struct, you need to read the field name as a string
    so you can find the struct field it is mapped to.
    Using `unsafe` will bypass the allocation and copying overhead of []byte->string conversion.

To install using unsafe, pass the 'unsafe' tag:

    go get -tags=unsafe github.com/ugorji/go/codec

Also, we cleaned up error handling so that an error implementation is always
used, and the error values could be shared if constant (i.e. use a var to keep
single instance instead of creating a new error value each time).

In addition, we now make error a method of Decoder and Encoder, setting the stage
for possibly moving away from panics to manage internal error propagation.
Note: I eventually decided not to do this (see comments in helper.go).