codecgen: do not add a separator until first entry is entered.
Previously, we entered a separator after the first entry is checked.
However, the first entry may be omitted (because it is the zero value).
Instead, keep track of when the first non-zero value is entered,
and only start entering a separator after that.

Also, for encode-struct-to-array, encode nil only if value is not empty.

Fixes #72