codecgen: Fix imports with same name e.g. "runtime" and "k2/runtime".
Do this by giving each import its own unique package name.

To make fix, keep a mapping of pkgpath to package name in file,
and use that when generating files.

For each package used in the file, which does not have a / in it,
use a unique package name.

For example, do:

import (
       "runtime"
       pkg1_runtime "k2/runtime"
       pkg2_ugorji "k2/ugorji"
)

Fixes #94