codec: inline anonymous fields if the struct tag is "effectively" blank.
Previously, we only inlined them if the struct tag was not present.
However, some folks have unknown flags in the struct tag. So it may be present
but effectively blank to us e.g. codec:",abc,def"
OR codec:",inline" which we see in Kubernetes.

For these, we handle gracefully.

Updates #95