codecgen: support extensions.
Previously, extensions were not supported in codecgen.
However, for completeness and to position codecgen as full replacement option,
we must support it, and try to make it as performant as possible.

This rounds up making sure that codecgen supports the same sequences that
the reflection-based encoding/decoding takes. The only missing feature is
support for Canonical=true option.

Updates #95