codec, codecgen: Support (M|Unm)arshalJSON exactly as encoding/json when using JSONHandle
To do this, we added support for explicitly tracking the bytes which have been read.
This way, when we see JSONUnmarshal, we can read the next set of bytes representing a value,
and pass that to the UnmarshalJSON method for custom decoding.

We also explicitly allow JSONMarshal to output JSON representation,
and expect it to be compliant.

This brings feature-parity with how JSON(M|Unm)arshal works,
which is very different from how encoding.Text(M|Unm)arshal works.

We also now only handle (M|Unm)arshalJSON when using JSONHandle,
and we check for it before encoding.Text(M|Unm)arshal.

This support applies to reflection-based and codecgen.

Fixes #97