codec: Selfer implementation should not call itself.
This effectively calls itself recursively and leads to an expected stackoverflow.
To fix, check if the variable name == the constant variable name of the type in the
Selfer implementation (x). If it is, then do not attempt to call the Selfer method.

In addition, the following fixes were applied:

When encoding extensions, use an address if the kind is not Struct or Array.
This way, we don't copy unnecessarily.

When encoding based on the known interfaces which may be implemented
i.e. encoding.(Binary|Text)(M|Unm)arshaler, json.(M|Unm)arshaler, codec.Selfer,
the value to encode may not be addressable but the implementation is on a pointer.
In that situation, copy the value to an addressable value and call the implementation there.

Updates #100