codecgen: Allow users specify random integer for use in generated code.
This allows for deterministic output during codecgen runs.
It also reduces the number of diff lines in the version control system (e.g. git).

Do this by allowing users specify a -d parameter to codecgen, which is
used in genRunner x.xs and used in the names of temporary generated files.

Fixes #101