codec(gen): code clean-up to remove separator concerns from framework and major re-architecture.
Up until now, the framework that handles encode and decode did all the work to manage
formats that used separators and those that didn't.

This led to the following issues:

- "framework" handling a lot of functionality that only applies to json
  e.g. handling separators [:,] appropriately.
- challenging to support other formats e.g. XML,
  which do not use separators but are text-based. It also made it hard to put changes in,
  since all work done by the framework in reflection-mode MUST be replicated in generated-mode.
- Generated code was long to handle all the ideosynchracies of the different formats.

We have now simplified the code to push handling of different separators and end tags to JSON
(who cares about it). This created a lot of cleanup by:
- removing separator handling code and its complexities from reflection and codecgen code
- reduced the size of generated files

JSONHandle will now handle reading and emitting separators for arrays, maps and closing tags
itself. It keeps a stack to track where it is in the "tree" and handles everything.

All these changes have been tested in all permutations, and they work.

In addition, we now keep track of number of bytes read (during decoding) and annotate
any error message with it, so we can track whee the decoding fails (if/when it does).

--------------

We also created a tests.sh script that runs all permutations of our functional
tests, running all tests agains reflection, codecgen and codecgen+unsafe,
and testing canonical, read/write from []byte or io.Reader|Writer, struct2array, etc.

All tests pass.

-----------------

In addition, we did the following:

- Binc encoding builtin for time.Time should handle *time.Time also.
- doc clear requirements for running python-based msgpack/cbor tests
- clearly document requirements for python-based msgpack/cbor tests.

-----------------

Finally, we updated GenVersion to 4, to reflect all these changes, and so
users have an quick fail and know to re-generate their static code.

-----------------

ENJOY.