codecgen: Support GO15VENDOREXPERIMENT
We depended on PkgPath to return the import path.
However, with GO15VENDOREXPERIMENT, reflect.Type.PkgPath()
will now return the path relative to the GOPATH, including /vendor/
in the path. This is no longer the import path.

We now will strip the vendor prefix appropriately if GO15VENDOREXPERIMENT is set.

We put a HACK comment in there, so we know to look at it later if things change
in the go sdk.

Fixes #102