codec: support max initial collection length; json fixes; improve perf & reduce alloc drastically
-------------
MaxInitLen Option:
It is possible to force an OutOfMemory crash when using the Decoder.
This is because we "make" a slice based on the pre-fix length in the stream
(e.g. for cbor, msgpack, etc).

To alleviate this, a user can configure a MaxInitLen, and we use that as the max
when making the initial collection (slice, map or chan). We then populate to that,
and do append's after that for any further items up to the length.

To use this, configure MaxInitLen in the Handle.

This support is exposed in runtime-reflection, fast-path and codecgen modes.

-------------
Growcap:
Add a smarter algorithm for append (which takes size class into consideration).
It also adds a simpler implementation for append, which doesn't use varargs
and just handles one append effectively.

-------------
ExpandSliceValue:
It also adds a ExpandSliceValue which does similar work to what
reflect.Append(...) does, but only the Grow part. This way, we use the smarter
append logic and do not require a slice of the varargs.

-------------
Performance: Reduce Allocation in encoding and decoding:
Previously, we wanted value methods, and kept the "optional" fields of
(en|de)cFnInfo in an embedded *(en|de)cFnInfoX struct.
This led to a lot of allocation.

However, we also store them in a slice field of the (En|De)coder.
It is thus easy to grab an internal pointer to them without doing an allocation.

To leverage this, we make all functions take a pointer receiver, and
just pass the internal pointer from the slice to them.

This reduced allocation by over 10% and improved performance.

-------------
sync.Pool for encoding structs while supporting omitEmpty:
Previously, the algorithm was incorrect. We incorrectly bypassed the Pool
about half of the time. The algorithm is now fixed.

-------------
Performance: JSON: reduce allocation using sync.Pool

JSON encoding/decoding uses a "stack" to manage the state as it transitions
in and out of arrays, maps/objects, etc.

The "stack" is implemented using a list. Making those lists leads to much allocation.

Since there's a clear start() and end(), we could use those to grab a slice from
the pool and return it when done parsing/encoding.

-------------
JSON: misc

Also, within DecodeBytes, if !isarray and zerocopy, try to use a scratch byte slice.

Also, remove some unnecessary code at the bottom of jsonStack.sep().
This reduced the function size and made sep inlinable.

Fixes $90