codec: support customization of struct tag keys; maintain default to "codec", "json".
To implement it, I will have to move from using a global variable for the typeInfos,
to using one that can be configured on the Handle.
If not configured, default to the current behaviour of looking at "codec" and "json" keys
(to support codec and migration from the standard library respectively).

This is implemented as follows:

- Define a TypeInfos which can be configured on the Handle
- If not defined, a default one will be used, which looks at codec, json keys
- codecgen will now take a parameter '-st',
  which will define a comma separated list of keys to inspect.
  It defaults to "codec,json".

The user will have to work around the possibility of being out-of-sync if using codecgen.
However, that is fine, because the user runs codecgen explicitly.

This will expand the API slightly:

- Introduce TypeInfos
- Add NewTypeInfos(tagKeys []string)
- Add TypeInfos field to BasicHandle, so every Handle can configure it
- Add -st parameter to codecgen

By making TypeInfos configurable from the Handle, a user can customize
the struct tag keys that are used when inspecting types.
This allows users to depend on custom keys e.g.

    type X struct {
      ID string `gorethinkdb:,omitempty`
    }

Fixes #105