codec: Fix UnmarshalJSON support
We changed the json decoding so it would reduce the number of unreads
and potentially keep track of the last byte read, which indicated
that the last json value was completely read.

However, this broke UnmarshalJSON support, which needed to use the
"cached" byte plus the subsequent bytes.

Fix this by adding an uncacheRead method to the decDriver, and letting
Decoder have a nextValueBytes method which is called from jsonUnmarshal.

Fixes #112