codecgen: Fix array decoding and generate code for Elem of seen reflect.Type
When we saw a reflect.Type, we recorded that a custom function for it
should be generated. However, we didn't record that possible Element types should
have generated code also. E.g. if we say [][2]int, we would generate
code for [][2]int, but not for [2]int, but the generated code for [][2]int
assumed that there was generated code for [2]int.

This is now fixed.

Also, the code generated for arrays kept some unused variables lurking.
This is now also fixed.

Fixes #118