codecgen: during encode, use address of value if a pointer receiver and a non-pointer value
When generating code for encode, we would have a variable which might be of pointer type
(if the kind was struct or array), else it was of non-pointer type.

However, we would always pass in the variable whether the type t or ptrTo(t)
implemented the known interfaces: (Binary|Text|JSON)(M|Unm)arshaler

Now, we pass in the address of the value if a pointer receiver and a non-pointer value correctly.

Fixes #120