Correctly decode deeply nested arrays (etc.)
When decoding "naked" data (to an arbitrary interface{}), and handling
nested maps and arrays, save a pointer to the slice slot around the
recursive call to decode().  Since the recursive call could call append()
to extend the slice, the underlying storage might move, resulting in a
case where we recursively decoded and wrote into the old array, but then
return a value from the new array.