codec: fix nested naked decoding, especially for lists and extensions
During naked decode, we use a list as a poor man's slab allocator.
When we need to use an item from the slice, we must keep a reference to it outside of the slice.
When nested decode happens, and an append causes a new slice to be created,
we will still be working with the same value.

Fixes #122