codec: json: support encoding large integers as a number or string
Previously, all integers were encoded as json numbers. However, some
large integers (|x| > 2^53) could not be encoded as
json numbers (IEEE double precision) without loss of precision.

To mitigate it, we now support options to define how integers are to be
encoded, either always as json numbers, always as json strings, or
as mostly as json numbers but as json string if |x| > 2^53.

On decode, we now always support decoding a number from a JSON string.

This allows users to ensure that precision is not lost.
It doesn't solve fully the issue where some other libraries some numbers
to be encoded as json strings and others as json numbers. We consciously will
not support it, treating it as an edge case and the library provides the tools
to support the compatibility already (e.g. use a string struct member, pass
the 'A' or 'L' option to IntegerAsString option, etc).

Tests are included.

Updates #128