codecgen: CLI: always assume that vendoring is enabled.
When generating the transient go file for running codecgen,
assume that vendoring is enabled.

This is typically true in go 1.6 and always true in go 1.7+.
When it is not true, we will be ok. Someone shouldn't run
codecgen on a package with vendor/ in its path, if vendoring is
not being done.

Fixes #164