Decode lists/arrays into arrays instead of slices.
This breaks usages where the length of the array isn't known initially;
the broken tests are commented out -- the RPC tests and a cbor test.

To accommodate cases like this, and also not take the performance hit if
this feature isn't wanted, decoding into array vs slice should be
configurable (like the map type is configurable). That's still to-do.