codec: json should encode regular floats without 'E' all the time.
We previously encoded floats using 'E' format, which meant that even
for a float like 3232, we encoded as 3.232E+3 (instead of 3232.0).

Now, we encode as 3232.0.

Note: For precision (as this package tries to infer whether a float or an int),
we have to always have an E or a . in the number when encoding.

Fixes #174