codec: perf: make common-path of bytesEncWriter.grow inlineable
Most requests to grow for a bytesEncWriter will either move
the cursor or expand the length to the cap of the []byte.

Put this in a different method that can be inlined, and only
call out to the growCap method if an allocation is necessary.

Makes encode for a []byte about 8% faster in microbenchmarks.