codec: remove recognized-rtid logic
Trying to see if a type is a recognized type
(one of the base go types or maps or slices containing them)
so that we can use a fast-path type-switch did not give the
performance we wanted.

We kept it gated behind a const flag (set to false),
but it is time to remove it for good now.