codec: error if need to expand non-settable slice, and denote whether decFn can accept ptr or value.
It is ok if we cannot reset the slice length, as long as the slice can be decoded into.
However, if we need to reallocate, and cannot reset, then throw an error

Also, map keys CANNOT be nil.
A nil value found in the encoded stream where a map
key is expected will trigger an error.

Use a flag (addrF) to determine whether the decode
function can accept either a value or a base type.