codec: support time.Time natively in all formats, and remove Builtin mode.
All codec's now have native support for time.Time, either as a builtin
type, an extension or a convention (e.g. json using RFC3339).

- json: RFC3339
- cbor: timestamp extension
- msgpack: timestamp extension
- binc: built-in time type
- simple: built-in time type

Consequently, it is ok to remove the "IsBuiltin" functionality
and assume all codecs support time.Time.

We also removed builtin functionality from taking effect.
All formats say that they do not support builtins.

However, to ensure that consuming libraries that have
codecgen'ed files do not need to re-generate, we allowed
the functionality as stubs so the files compile but the
code path is never run.