codec: test: move AnonInTestStrucIntf from values_test.go into values_flex_test.go
Some external libs (e.g. easyjson, ffjson) cannot initialize a pointer field
which is nil, making it impossible to run benchmarks against them.

So we just moved it into values_flex_test.go, so we still validate it for codec,
but can run with the highest common denominator for benchmarks.