codec: streamline panic/recover error handling, manage compatibility with codecgen and support structs where fields are encoded as numbers
Add diagnostic information (encode format, num bytes read, etc) to error on panic-recover.
This required the following changes:
- errors are either exposed as a string or an error, and used as the panic value
- on recover (in Encoder.Encode or Decoder.Decode), we add diagnostics to the panic value and return as error

encDriver for a handle should not check for overflow32 - instead, just expose a DecodeFloat64() method

We currently keep some deprecated methods alive in (en|de)cDriver, so that codecgen will not need to be re-done.
Instead, for compatibility in codecgen, make the r variable a struct wrapping (en|de)cDriver.
Thereafter, add the compatibility functions, so they can be removed from (en|de)cDriver.

Make clear that float overflow doesn't mean that the number maintains its precision.

In JSON, always encode and decode as float64.
If not, we lose precision when we encode as float32 and decode back as float64 (e.g. for math.MaxFloat32).

Support encoding/decoding structs where the encoded stream should use number codes for the field names (instead of strings).
This is now supported by configuring the key type on the _struct field,
and supported in codecgen and in the reflection-based general support.
See documentation on (*Encoder).Encode.