codec: refactor struct field loading and data structures, for performance and memory friendliness
Instead of using a structure, and checking if strings match constantly,
we store all the field names in a byte array, and just use a bytes.Index
check to see if any field name has been seen before.

This saves on the size of typeInfoLoad while making this more streamlined.