codec: remove double-indirection and fix race in Copy-On-Write during TypeInfo lookup
Previously, we stored a pointer to a slice as an atomic.Value.
This meant that we were effectively storing a pointer to a pointer to an array.
Now, we just store the pointer to the array and the length separately,
and "compose" the slice using unsafe (in unsafe mode).

This gave a slight perforamce increase.

Also, we now create a new slice whenever we need to update the TypeInfo array.
Previously, we just updated the backing array and returned a new slice object.
But since the same backing array was used, it introduced a race condition.

Fixes #227