codec: streamline chan encoding/decoding
Previously, we would encode a chan by reading up to the length of the chan.
This is problematic because
- we do not consume any elements currently blocking due to a smaller chan cap
- we have no option to drain the chan until it is closed.

To support all models, we need to be able to collate all values received from the chan,
decipher the true number of values received, and then encode the true length and values appropriately.
This suggests that we should receive the values into a temporary slice, and encode that slice.

This is now supported fully, across standard reflection-based path and the codecgen pathway.
It is configured via the ChanRecvTimeout parameter in EncodeOptions, available to all handles.

TestStrucFlex was also expanded, to test and validate the expanded chan support.

Fixes #233