Implement 'omitemptyrecursive' tag (#10)
This is in spirit a reimplementation of a previous PR that implemented an
'omitemptycheckstruct' tag. This is necessary because, even though we
already have a RecursiveEmptyCheck field in EncoderOption, flipping it
on results in undesired behavior, like existing pointer fields with omitempty
being checked recursively, and not being able to specify otherwise.

This new 'omitemptyrecursive' tag is a combination of 'omitempty' and a
per-field version of RecursiveEmptyCheck.