codec: support ability to preserve missing fields
Given multiple versions of a type, where some fields are removed,
we want to support the ability for the encoding stream to remain canonical.
This means that the value which encoded to a stream of bytes before
can be recovered into a different version of the type.

This is supported by implementing the MissingFielder interface.

Note that the MissingFielder interface is completely ignored during codecgen.

Updates #258