codecgen: add support for go1.11 modules
Without this change, when using the new module support in go 1.11, codecgen would not be able to determine the proper package path and fail.

Since the golang.org/x/tools/go/packages does not yet support previous go versions, use a build tag to revert to the previous mechanism on older go versions.