codec: provide TimeNotBuiltin option to allow time.Time be handled in custom way
Previously, time.Time was handled either in a standard way (based on the format)
or in in a custom way that defaulted to using a custom extension,
or allowing the encoding.(Binary|Text)(M|Unm)arshaler implementation to be used.

Around Nov 2017, that was changed, as all formats provide a standard way or a
standardized extension format for encoding time.Time.

The option: TimeNotBuiltin, now allows users the option to customize how time.Time
is handled (as it was before).

Fixes #269