codec: limit size of bytes read to MaxInitLen (specifically for extensions)
Previously, we called readx for extensions. However, readx will, as needed,
make a slice with the capacity passed in. A bad length of the extensions
bytes can cause a request for very large memory.

To mitigate this, we now call readx only for bytesDecReaders, else we
delegate to decByteSlice function which is designed to incrementally
create the []byte as needed.