codec: limit depth of nested structures
Currently, codec will happily decode your value up to any depth.

This can lead to a DOS, when bad input is fed into the engine, causing
a stack overflow error or other possibly fatal memory issues.

We now mitigate against this, by allowing a configurable max depth.
To ensure that older users are not bitten by this, the max depth
currently defaults to 1024. Users can configure it accordingly.

With this fix, we carefully only do the checks when decoding into a slice/array,
a map or a struct (kMap, kSlice and kStruct methods of *Decoder).
We also integrate these checks directly into the fast-path functions.

Kudos to @akalin-keybase for discovering this issue and proposing an initial fix.