codec: rpc: use lock free mechanisms to track the close status.
Rpc needs to check whether the connection is closed everytime a decode is to be done.

However, the closing can happen in a different goroutine.

Previously, we used a mutex to write (Lock/Unlock) and read (RLock, RUnlock).

However, this can be expensive, since the writing happens at most once, but the
reading happens everytime.

We now make this more performant, by using lock-free mechanisms
(atomic.Value and atomic.Load/StorePointer).

This may improve some scalability and stress scenarios.

Fixes #271