codec: implicitly call Close() by default
We introduced a Close() method to Encoder and Decoder.
This gives users a way to explicitly close and returned shared resources
- [N]byte buffers used by bufioEncWriter and bufioDecReader
- large structures needed for naked decoding effectively
- structures needed for mapping types to encode/decode functions

However, forcing users to explicitly call Close() is unexpected, and
should not be the default.

We now have a flag: DoNotClose (default=false). If set to true, then
we will not Close() after a Encode/Decode call.

As extensions can call Encode/Decode within a corresponding Encode/Decode
call, we use a counter to increment each time Encode/Decode is called.
When that counter resets to 0, we will call Close() unless DoNotClose=true.

In addition, we ensure that tests have DoNotClose=true configured, and we
handle explicitly calling Close() within the tests.