codec: at start of encode, ensure bufioEncWriter's buffer is not nil
We previously introduces a mechanism where calling Release was optional,
and we would implicitly call Release if a full Encode was completed.

Unfortunately, this broke the use-case where an Encoder was used
multiple times in sequence.

We fix this now, by doing what we should have done before i.e. in addition
to calling Release (which nullifies the []byte buffer and returns it to the
pool), we should also grab a buffer from the pool at the beginning of each
Encode call.

Fixes #278