codec: introduce EncodeOptions:StringToRaw flag to control string encoding
Previously, we based the implementation on the idea that string types are always UTF8.
That is incorrect.

In go, a string is just an immutable sequence of bytes (i.e. immutable view of []byte).

Now, we honor that, and provide an EncodeOptions flag called StringToRaw that configures
how we encode strings - either as a "raw" sequence of bytes, or a UTF-8 string.

This is supported in reflection-based path, fast-path and codecgen.

While at it, remove the loop-unrolling done in fast-path.
This makes the code easier to read and we don't have to modify stuff in multiple places.
We will let the compiler and CPU branch predictor help remove the performance impact of
the unrolled loops, which is easier as the condition is constant throughout the loop.

Fixes #286