codec: msgpack: Honor RawToString when decoding raw bytes into empty interface{}
RawToString says that we can see raw bytes in a stream,
and decode it as string (as opposed to bytes).

There was a situation where we didn't honor that.
When we saw raw bytes but we were using the old spec,
we blindly said that it had to be []byte. However, the
RawToString is there to configure that behaviour.

I need to make time to do a comprehensive set of tests on
RawToString and StringToRaw. That is coming.

Fixes #293