codec: do not generate fast-path for uncommon types, and generate sort.Sort wrappers
We no longer generate fast-paths for maps with
- keys of type int8/16/32, uint16/32, float32/64, bool, interface{}
- values of type int8/16/32, uint16/32

This is in a bid to reduce the binary size for types which depend on go-codec.

Previously, go-codec was adding about 11MB to a binary.
We have now reduce that to 5.7MB.

The user has the option to reduce this further to
2.9MB by passing the tag 'notfastpath' during a build or install.

In addition, we now generate sort.Sort wrappers for types we care about.
This way, we don't have to maintain the code manually.

In generated fast-path and other code, we also reduced the amount of unnecessary
conversions we do.

Updates #295
Updates cisco/senml#22
Fixes #296