codecgen: use fully qualified name even for embedded fields
This allows us support embedded types with similarly named fields.

For example, if a struct T has a field named T, and the type T is
embedded within another type, then previously we used the shorthand
of the fieldname (i.e. instead of x.T.T, we used x.T) which was
resolved to the embedded type not its embedded field.

Now, we always use the fully qualified name.

Fixes #297