codec: skip subsequent structFieldInfo with same encName and same depth
We missed a case where we see a structFieldInfo with same encName and same depth.
We should have skipped it but didn't.

This occurs if you embed 2 types in a struct, and both of them have a field with
the same name at the same level.

Fixed now, with a test added.

Fixes #297