codec: fast-path: optimize: remove unnecessary local variables
rearrange structs so that commonly accessed fields are at top of any cache lines,
and within 1st 2 cache lines of the enclosing struct.

Do not copy encDriver, decDriver, esep into local variables.
Instead, directly use the field in *Decoder or *Encoder, as these are performant
to access (since registers are not used for local variables).