codec: json: try fast-path for decoding floats before defaulting to strconv.ParseFloat
The json format for floats is more constrained that what strconv supports.
We noticed some slowdown with float parsing which we have attempted to resolve
by walking through the text and doing an exact parse if within the narrow constraints where
it is simple.

Also, have float32 not convert back and forth between float64 unnecessarily.