codec: handle []byte encoded as array (as opposed to native mode per format)
Generally, formats (e.g. cbor, msgpack, etc) have a defined way of encoding
binary data ([]byte). However they also support encoding any array of values.

It is possible to encode []int8 of positive numbers, and then decode that
into []uint8. In this mode, the decoded stream doesn't look like the native
bytes in the stream, but just like a sequence of small positive numbers.

We supported this before in a non-performant way. Now, we support it better
by having the driver fully read a sequence of numbers.

While there, do not create special methods for []uint8 or []uintptr, or
maps with uintptr keys or values. uintptr is uncommon in use, so was cut
off to manage binary size.

Finally, expand the values in generated tests so we exercise more float types.